src/requirements.txt
requests==2.28.2
azure-identity==1.19.0
pylint==2.17.2
openai==1.52.0
azure-cognitiveservices-search-newssearch==2.0.0
python-dotenv==1.0.1
yfinance=0.2.46
semantic-kernel=1.11.0